moving useTips hook to the api module
